Move PrintAsID into explorer.
