+++ explorer/ast/BUILD
@@ -57,6 +57,7 @@ cc_library(
         "//explorer/base:decompose",
         "//explorer/base:error_builders",
         "//explorer/base:nonnull",
+        "//explorer/base:print_as_id",
         "//explorer/base:source_location",
         "@llvm-project//llvm:Support",
     ],
@@ -153,6 +154,7 @@ cc_library(
         "//common:ostream",
         "//explorer/base:error_builders",
         "//explorer/base:nonnull",
+        "//explorer/base:print_as_id",
         "//explorer/base:source_location",
         "//explorer/base:trace_stream",
         "@llvm-project//llvm:Support",
