+++ explorer/interpreter/BUILD
@@ -25,6 +25,7 @@ cc_library(
         "//explorer/base:arena",
         "//explorer/base:error_builders",
         "//explorer/base:nonnull",
+        "//explorer/base:print_as_id",
         "//explorer/base:source_location",
         "@llvm-project//llvm:Support",
     ],
@@ -129,6 +130,7 @@ cc_library(
         "//explorer/ast:expression_category",
         "//explorer/base:arena",
         "//explorer/base:error_builders",
+        "//explorer/base:print_as_id",
         "//explorer/base:source_location",
         "//explorer/base:trace_stream",
         "@llvm-project//llvm:Support",
@@ -144,6 +146,7 @@ cc_library(
         "//explorer/ast",
         "//explorer/base:error_builders",
         "//explorer/base:nonnull",
+        "//explorer/base:print_as_id",
         "//explorer/base:trace_stream",
         "@llvm-project//llvm:Support",
     ],
@@ -159,6 +162,7 @@ cc_library(
         "//explorer/ast",
         "//explorer/ast:static_scope",
         "//explorer/base:arena",
+        "//explorer/base:print_as_id",
         "//explorer/base:trace_stream",
         "@llvm-project//llvm:Support",
     ],
@@ -221,6 +225,7 @@ cc_library(
         "//explorer/base:arena",
         "//explorer/base:error_builders",
         "//explorer/base:nonnull",
+        "//explorer/base:print_as_id",
         "//explorer/base:source_location",
         "//explorer/base:trace_stream",
         "@llvm-project//llvm:Support",
@@ -242,6 +247,7 @@ cc_library(
         "//explorer/ast:static_scope",
         "//explorer/base:error_builders",
         "//explorer/base:nonnull",
+        "//explorer/base:print_as_id",
         "//explorer/base:trace_stream",
         "@llvm-project//llvm:Support",
     ],
